Create dir utils & rename files
Moved GeneradorID & CapturarDatos to utils. Renamed function files from camelCase to kebab-case
