+++ src/utils/capturar-datos.ts
